fix: update license-flow page
